imunemais-api/pyproject.toml
[project]
name = "imunemais-api"
version = "0.1.0"
description = ""
authors = [
    {name = "DevChimas1",email = "devchimas@gmail.com"},
    {name = "Jorge Gabriel Rodrigues Ferreira",email = "jorgegabrielrf@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.11.5,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "imunemais_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
